Refactor maintenance into own module
